Move HTTPRedirect definition to Rover
