+++ kifi-backend/modules/common/app/com/keepit/rover/fetcher/HttpRedirect.scala
@@ -1,6 +1,7 @@
-package com.keepit.scraper
+package com.keepit.rover.fetcher
 
 import com.keepit.common.net.URI
+import com.keepit.scraper.ShortenedUrls
 import play.api.http.Status._
 
 case class HttpRedirect(statusCode: Int, currentLocation: String, newDestination: String) {
