+++ kifi-backend/modules/scraper/app/com/keepit/scraper/ScrapeWorker.scala
@@ -6,6 +6,7 @@ import com.keepit.common.logging.Logging
 import com.keepit.common.healthcheck.AirbrakeNotifier
 import com.keepit.common.net.URI._
 import com.keepit.model._
+import com.keepit.rover.fetcher.HttpRedirect
 import com.keepit.scraper.extractor._
 import com.keepit.scraper.fetcher.HttpFetcher
 import com.keepit.search.{ LangDetector, Article, ArticleStore }
