+++ kifi-backend/modules/scraper/app/com/keepit/scraper/fetcher/HttpFetcher.scala
@@ -2,7 +2,8 @@ package com.keepit.scraper.fetcher
 
 import com.keepit.common.net.URI
 import com.keepit.model.HttpProxy
-import com.keepit.scraper.{ HttpRedirect, HttpInputStream }
+import com.keepit.rover.fetcher.HttpRedirect
+import com.keepit.scraper.HttpInputStream
 import org.joda.time.DateTime
 
 import scala.concurrent.Future
