+++ kifi-backend/modules/scraper/app/com/keepit/scraper/fetcher/apache/FetcherHttpContextAdaptor.scala
@@ -1,8 +1,8 @@
 package com.keepit.scraper.fetcher.apache
 
+import com.keepit.rover.fetcher.HttpRedirect
 import com.keepit.scraper.fetcher.FetcherHttpContext
 import org.apache.http.protocol.{ HttpContext => ApacheHttpContext }
-import com.keepit.scraper.HttpRedirect
 
 private[apache] class FetcherHttpContextAdaptor(context: ApacheHttpContext) extends FetcherHttpContext {
   def destinationUrl: Option[String] = Option(context.getAttribute("scraper_destination_url").asInstanceOf[String])
