+++ kifi-backend/modules/scraper/test/com/keepit/scraper/fetcher/FakeHttpFetcher.scala
@@ -3,7 +3,8 @@ package com.keepit.scraper.fetcher
 import com.keepit.common.akka.SafeFuture
 import com.keepit.common.net.URI
 import com.keepit.model.HttpProxy
-import com.keepit.scraper.{ HttpRedirect, HttpInputStream }
+import com.keepit.rover.fetcher.HttpRedirect
+import com.keepit.scraper.HttpInputStream
 import org.joda.time.DateTime
 import play.api.http.Status
